Fisrt commit - GPIO basic - blink led
In this fist commit I sent the folder with the blink led ESP-ODF program (C language) as and example of the GPIO as output. It uses the freeeRTOS to make a delay (vTaskDelay function) and the GPIO set of functions to reset, configure and send data through a specific IO pin.