GPIO RGB led blink using array
This commit contains the essential files to overlay a project based on ESP32(s2) and thus be able to build and flash boards of this type of SOC. The objective is to flash the three colors of the RGB LED mounted on the board. Pinning in program description (in main.c).